Add Source Map spec to the list
Add the latest revision of the Source Map spec to the list, using the readable
HTML version of the spec published by sourcemaps.info. The actual spec is in a
Google Doc but there is no easy way to extract content automatically from a
canvas-based Google Doc:
https://docs.google.com/document/d/1U1RGAehQwRypUTovF1KRlpiOFze0b-_2gc6fAH0KY0k

Organization and group are set to "sourcemaps.info" for lack of better choice.
Also note the need to relax a couple of tests as there is no real repository
for the spec (there is one for sourcemaps.info, but that's for the site itself,
not for raising issues against the spec).

Closes #618.

This will add the following entry to the list:

```json
{
  "url": "https://sourcemaps.info/spec.html",
  "seriesComposition": "full",
  "shortname": "sourcemap",
  "series": {
    "shortname": "sourcemap",
    "currentSpecification": "sourcemap",
    "title": "Source Map Revision 3 Proposal",
    "shortTitle": "Source Map",
    "nightlyUrl": "https://sourcemaps.info/spec.html"
  },
  "shortTitle": "Source Map",
  "organization": "sourcemaps.info",
  "groups": [
    {
      "name": "sourcemaps.info",
      "url": "https://sourcemaps.info/"
    }
  ],
  "nightly": {
    "url": "https://sourcemaps.info/spec.html",
    "filename": "spec.html"
  },
  "title": "Source Map Revision 3 Proposal",
  "source": "spec",
  "categories": [
    "browser"
  ]
}
```