Set --no-cone mode for git sparse-checkout of WPT (#661)
Behavior of sparse-checkout is not fully stable and changed in 2.37 version of
Git. Command line now expects directories by default, and not patterns. Forcing
the mode to `--no-cone` tells Git to parse the parameter as a pattern instead:
https://www.git-scm.com/docs/git-sparse-checkout#_internalsfull_pattern_set

That is what we need, although note use of the `--no-cone` feature is
discouraged and may be remove in a future version of Git. Feature does what we
need for now. In the absence of it, we'd have to build the list of directories
and `META.yml` files ourselves. That's doable with the right Git command and
parameters, just requires looking into what the right Git command and
parameters are ;)