Add a new alternateUrls property and use it for CSS drafts
This records the `w3c.github.io` URL for CSS drafts in a new `alternateUrls`
property under `nightly`. See context in #704.

That property is an array as the mechanism could be used to record nightly URLs
used by a spec throughout its development. The code automatically adds the
w3c.github.io URL for CSS drafts.

One possible question: A spec already published in /TR could get a new
shortname. Since the array is under `nightly`, this cannot really be used to
record the previous /TR URL. Should this `alternateUrls` property rather be
moved to the root (but then how to distinguish between alternate URLs for the
nightly and release versions of the spec)? Or would we add an `alternateUrls`
property under `release` as well if we ever want to track that?