Add a new alternateUrls property and use it for CSS drafts (#706)
This records the `w3c.github.io` URL for CSS drafts in a new `alternateUrls`
property under `nightly`. See context in #704.

That property is an array as the mechanism could be used to record nightly URLs
used by a spec throughout its development. The code automatically adds the
w3c.github.io URL for CSS drafts.