Add FIDO CTAP spec to the list (#722)
This adds the FIDO Client to Authenticator Protocol spec to the list, following
discussions in #708.

The latest revision of the spec seems to be the Proposed Standard. This adds the
version that includes errata.

Organization and group info are explicitly provided in `specs.json`. We may
consider custom logic when and if we add more FIDO specs.

The shortname used by the FIDO Alliance uses a slightly different convention
than other shortnames, adding a `v` before the version number: `fido-v2.1`. The
shortname schema had to be relaxed accordingly.

Similarly, the filename schema had to be relaxed to aslo accept `.` characters.

The spec links to a GitHub repository that does not exist, or more probably that
isn't public, so no repository.

This would add the following entry:

```json
{
  "url": "https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-errata-20220621.html",
  "seriesComposition": "full",
  "shortname": "fido-v2.1",
  "series": {
    "shortname": "fido",
    "currentSpecification": "fido-v2.1",
    "title": "Client to Authenticator Protocol (CTAP)",
    "shortTitle": "CTAP",
    "nightlyUrl": "https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-errata-20220621.html"
  },
  "seriesVersion": "2.1",
  "organization": "FIDO Alliance",
  "groups": [
    {
      "name": "FIDO Alliance",
      "url": "https://fidoalliance.org/"
    }
  ],
  "nightly": {
    "url": "https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-errata-20220621.html",
    "alternateUrls": [],
    "filename": "fido-client-to-authenticator-protocol-v2.1-ps-errata-20220621.html"
  },
  "title": "Client to Authenticator Protocol (CTAP)",
  "source": "spec",
  "shortTitle": "CTAP",
  "categories": [
    "browser"
  ]
}
```

Co-authored-by: Dominique Hazael-Massieux <dom@w3.org>