Complete alternateUrls with series URL for CSS drafts (#725)
This adds the alternate series URL to the `alternateUrls` list for CSS drafts.

The alternate series URL is only added when the entry under consideration is the
current level in the series. This guarantees that an alternate series URL
appears only once in the resulting list.

This fixes #720 by implementing option 3 discussed there: "Add the series
alternate URL directly to `nightly.alternateUrls`".