Test unicity of sourcePath and fix duplication in SVG11 entry (#726)
This adds a test to ensure that we cannot have two different nightly versions of
a spec that target the same repo and source path.

This typically happened with SVG11 and SVG2. The SVG11 entry incorrectly
reported that its source is to be found in the w3c/svgwg repository. This update
drops the association with that repository... and updates the underlying code
in `compute-repository` not to take for granted that hypothetical repositories
returned by `parseSpecUrl` actually exist.

This fixes #418.