Use draft shortname for alternateUrls of CSS specs
The `alternateUrls` field for CSS specs was computed from the spec's shortname.
There are a few cases where the draft shortname is not the TR shortname,
e.g. `css-device-adapt` vs. `css-viewport`.

This update also hardcodes the exception to the rule for the `css` series
shortname, which CSS uses for CSS snapshots and not for the CSS 2.x series.