Add https://www.w3.org/TR/test-methodology/ (#775)
This document is eligible under criterion 4, specifically "*it contains
informative content that other specs often need to refer to (e.g. guidelines
from horizontal activities such as accessibility, internationalization, privacy
and security)*."

It contains the definition of "testable assertion" in particular.

Note the spec does not follow the more recent data definitions model.
Previous versions of Reffy would flag the dfns as being "private" as a result.
New version will detect that situation and export the dfns so that dfns can
be referenced through ReSpec xref.