Force series title for Screen Orientation (#794)
* Force series title for Screen Orientation

The title of the spec was recently changed from "The Screen Orientation API" to
"Screen Orientation". The series' title has not yet been updated in the W3C API
though. This update updates the series title in browser-specs while this gets
fixed in the W3C API.

* Add series shortname for Screen Orientation

Shortname is useless in this case but required per schema, because schema is
shared between `specs.json` and `index.json`. We should consider creating a
dedicated schame for `specs.json` that would not require `shortname`.