Add promises-guide, make code handle TAG findings (#806)
* Add promises-guide, make code handle TAG findings

This adds the "Writing Promise-Using Specifications" TAG finding to the
(non-browser) list to reduce the gap between specs in Shepherd and specs in
Webref, as discussed in https://github.com/tabatkins/bikeshed/issues/2431

The TAG was added to the list of non-browser groups to make sure TAG Findings
get classified as non-browser specs.

This also adds the code needed to extract findings shortnames and repositories,
so that we can be more systematic about these specs from now on.

The update also fixes the code that looked for a non-browser group in the list
of groups that produced the spec. That code assumed that whenever there's a
non-browser group in the list, the spec is not targeted at browsers.
CSS Typed OM is a counter-example (jointy developed by the CSS WG and the TAG).
The fix should not affect any other spec (all other specs jointly developed by
more than one group are developed by working groups targeting browsers).

* Adjust tests on categories

"Default" does not mean anything, there's always a group defined. Also added a
test on the new rule.

* Adjust unknown URL pattern test

Bad idea to have used TAG findings as an example of an unknown pattern ;)