Add new "standing" property
Second take at #812.

This update creates a new `standing` category for specs whose status is
"A Collection of Interesting Ideas" or "Unofficial Proposal Draft".

As usual, default rule can be overridden in `specs.json`. Done for
font-metrics-api-1 and web-animations-2 because they define IDL that has already
been included in Web Platform Tests.

The css-typed-om-2 spec will be the only one to get a `"pending"` standing for
now.

Specs in pending status appear in the web-specs package but get filtered out of
the browser-specs package.

The new status will make it possible to add a few additional unofficial CSS
specs. To be done in a separate commit.