Retrieve W3C API key from environment, drop config.json (#815)
There was a typo in the tests that checked groups: the code read the GitHub
token in `config.json` under `githubToken` instead of under `GH_TOKEN` as done
in `build-index.js`.

This meant the test request to the GitHub API was done without any
authentication. Most of the time, that did not have any consequence, but CI
tests now regularly fail at that step, probably because the GitHub API enforces
limits more strongly.

In any case, it seems much much better to use the GitHub token that GitHub
provides to actions rather than one we set by hand long time ago. I suspect
that setting config.json as a secret was done at a time when we were not
familiar with GitHub actions and secrets.

This update drops the `CONFIG_JSON` secret altogether and updates the code to
rather rely on the `GH_TOKEN` and `W3C_API_KEY` environment variables to read
the tokens it needs to build the index and run the tests.

The update also fixes the typo, of course.