Add "discontinued" standing, add missing specs for Bikeshed (#817)
This update adds a new `discontinued` value to the `standing` property for specs
that are still in the list for some reason (typically because other specs still
reference terms and sections they define in informative prose), but whose
development has been discontinued. Specs whose status is "Discontinued Draft"
end up with a `discontinued` standing.

This affects DNT in the current (before this update) list. The spec will be
filtered out of the `browser-specs` package as a result. That seems a good
thing.

This update also adds CSS specs that are needed for cross-referencing purpose in
Bikeshed, as requested in:
https://github.com/tabatkins/bikeshed/issues/2431#issuecomment-1373992354

New specs are either:
- discontinued specs such as selectors-nonelement-1, DOM-Level-2-Style
- older versions of some specs such as css-color-3, css-contain-1, css-ui-3
- unofficial drafts that have not yet been completely adopted by the CSS WG such
as css-color-6, css-color-hdr, css-conditional-values-1
- the latest CSS snapshot css-2022.