Stop re-building the index on pushes to the main branch
Building the index when we push a change to the main branch is a nice idea in
theory, but the build process consumes a lot of GitHub API requests and, due to
GitHub API rate limits, can essentially only run once per hour. That does not
work well when there is more than one pull request to merge.

The job can still be triggered manually through GitHub's admin interface.