Make compositing-2 current, fix series.releaseUrl logic
This makes `compositing-2` the current specification for `compositing`, as
discussed in https://github.com/w3c/browser-specs/pull/848#issuecomment-1403172297

This will make the data report the following series info:

```json
{
  "shortname": "compositing",
  "currentSpecification": "compositing-2",
  "title": "Compositing and Blending",
  "shortTitle": "Compositing",
  "nightlyUrl": "https://drafts.fxtf.org/compositing/",
  "releaseUrl": "https://www.w3.org/TR/compositing/"
}
```

The `nightlyUrl` is good and returns Level 2. The `releaseUrl` returns the
latest published version of the spec, which is Level 1. That's not the level
that is deemed current, but then it is the latest published version.

README amended accordingly to clarify that `releaseUrl` is not always the
current spec, and that `nightlyUrl` should always be the current spec (unless
there's a hiccup somewhere because /TR disagrees with the working group).