Add RDF 1.2 and SPARQL 1.2 drafts
This update adds 19 RDF/SPARQL 1.2 drafts to the list, as non-browser specs. It
also make the code cognizant of the RDF/SPARQL numbering and titling scheme.

Note that the shortname are "invented" for now. They match those used within
the specs themselves, so hopefully the working group will just continue to use
that same pattern when it starts publishing specs to /TR:
https://w3c.github.io/rdf-primer/spec/#related

Via #855.