Add RDF 1.2 and SPARQL 1.2 drafts (#864)
This update adds 19 RDF/SPARQL 1.2 drafts to the list, as non-browser specs. It
also makes the code cognizant of the RDF/SPARQL numbering and titling scheme.

Note that the shortnames are "invented" for now. They match those used within
the specs themselves, so hopefully the working group will just continue to use
that same pattern when it starts publishing specs to /TR:
https://w3c.github.io/rdf-primer/spec/#related

Previous RDF 1.1 entries (concepts and semantics) were dropped in favor of the
new ones. This means that the list does not contain any /TR document for RDF
and SPARQL, but that seems fine to me. Previous ones did not respect the
definitions data model and are thus "less interesting" in any case.

This update does not add the "What's new" and "Primer" specs (two for RDF, one
for SPARQL) that will be published as Notes.

Via #855.