Force status of Web Animations 2
The spec is being staged for publication to /TR and the status of the nightly
version is currently "First Public Working Draft". I suspect the nightly draft
will become an "Editor's Draft" after publication.

In the meantime, to unlock build, this update forces the status to what it used
to be, meaning "Unofficial Proposal Draft".