Temporarily force series title for scroll-animations (#921)
* Temporarily force series title for scroll-animations

The title of the spec switched to "Scroll-driven Animations" but the series
title in the W3C API is still the old one. Name will soon be fixed in the W3C
API database. This temporary update is only meant to resume building so that
a new package may be published.

* Also add series shortname

Needed as the schema for `specs.json` does not envision a `series` object
without it.