Handle false positives reported by find-specs (#956)
* Try to handle homepage links to explainer

The Storage Buckets repository has its homepage URL that targets the explainer
instead of the spec. This update attempts to detect that and reconstructs the
URL of the spec, not to report a false positive. Certainly not perfect, but that
works for now.

* Adjust monitoring key for Element Capture

The WICG repository's homepage URL links to the screen-share spec. This made
the find-specs script report a false positive because the monitoring entry was
for the `screen-share/element-capture` repo, and not for the spec URL itself.