Prefer index.html over Overview.html when detecting filenames
This update switches the order by which the code looks for spec filenames.

This follows from discussions about CSS drafts, which currently exist both as
`Overview.html` and `index.html`, the latter being the most recent:
https://github.com/w3c/csswg-drafts/issues/8899

This will de facto update the `filename` of all CSS drafts to `index.html`.
Note that this will also update the `filename` of WebCrypto, which is a good
thing because `Overview.html` actually redirects to `index.html` there.