Add step-by-step build mode with intermediary files
Build of the index file sometimes fails at intermediary steps and often fails at
the testing phase. In both cases, a new build needs to be run locally to
investigate more precisely what fails. That's tedious and takes time.

This update creates a "step-by-step" mode whereby the build records the result
of each step to an intermediary file in the `buildsteps` folder. These files
are only ever going to be useful for debugging purpose, but that's already a
good thing.

The build job now pushes intermediary files that got updated to the repository
no matter what, even when the build fails. This allows to restart the build
locally from the first step that failed.

The `10-filename.json` file in the `buildsteps` folder contains the final
results of the build (and thus matches the `index.json` when tests pass). That
file should be particularly useful to spot what needs fixing for tests to pass.

A possible alternative would have been to use separate branches but that seemed
much more complicated to setup and use.