New `obsoletedBy` and `formerNames` properties
The new optional `obsoletedBy` property makes it possible to identify specs that
replace a given discontinued spec in the list, as discussed in #1006. The term
`obsoletedBy` was preferred over `replacedBy` because the latter is not
fantastic to convery the integration of a spec into another one, and because
Specref uses that term.

The new optional `formerNames` property makes it possible to list former
shortnames of a spec in browser-specs.

Both properties are useful, and needed, for browser-specs to act as a source of
specs for Specref. Other projects may also use these properties to track a spec
over time.

Both properties must be set in `specs.json` and are copied over to `index.json`.

This update sets the `obsoletedBy` property on the discontinued Close Watcher
API spec to target the HTML spec. The `formerNames` property is not set on any
spec for the time being.

This update also re-introduces the Resizable ArrayBuffer proposal as a
discontinued spec obsoleted by ECMAScript. This is again needed for Specref, as
deleting the spec from browser-specs broke Specref's build, see:
  https://github.com/tobie/specref/issues/752