Add step-by-step build mode with intermediary files (#1013)
Build of the index file sometimes fails at intermediary steps and often fails at
the testing phase. In both cases, a new build needs to be run locally to
investigate more precisely what fails. That's tedious and takes time.

This update turns the build into a "step-by-step" build whereby the intermediary
files get created in a `.buildsteps` folder. These files are only ever going to be useful
for debugging purpose, but that's already a good thing.

The build job now stores intermediary files as artefacts on GitHub. These files
should be kept 90 days.

The `10-filename.json` file in the `.buildsteps` folder contains the final
results of the build (and thus matches the `index.json` when tests pass). That
file should be particularly useful to spot what needs fixing for tests to pass.