Explicitly ignore execCommand and sourcing in-band track
Both specs are referenced by HTML. Both specs are in a sorrow state with no
plan to resume work in the foreseeable future. Adding them to the ignore
list as a result.

We may want to revisit that if the situation evolves, but I don't think we
should (or want ;)) to do that every 3 months.