Fix spec loading with Puppeteer
Use of the `networkidle0` option does not work well with pages that maintain
"stream" connections to some HTTP endpoint, as done in some CSS drafts. This
update copies over and simplifies code from Reffy to ignore network requests
to such stream resources (and also to avoid downloading images since we don't
do anything with them).

Note I did not try to extract Reffy's code in the end, because it is a bit
more embedded in Reffy's logic and, e.g., also handles multipage specs, which
we're not interested in here.