Automate processing of IETF (group) drafts
The code now recognizes IETF draft documents that have a `datatracker.ietf.org`
URL:
- It associates them with the IETF organization
- It can compute a useful shortname (that code can in theory return a truncated
shortname because there is no direct way to validate that the Internet Draft
name contains a group ID).
- It extracts the group's ID from the nightly URL (that code could further be
improved to fetch the actual group name, right now the code only knows about
the "HTTP" working group).
- It associates IETF documents from the HTTP WG to the right repository.
- It computes the better-looking nightly URL at `www.ietf.org` or at
`httpwg.org` for HTTP WG documents.

This allows to simplify IETF data in `specs.json` a bit.

Note that the code still cannot process drafts that have been submitted by
individuals automatically, even when these drafts at targeted at a group.
Such drafts should be associated with the individuals that submitted them and
not with any group. A couple of spec entries, which incorrectly referenced the
Network WG or the HTTP WG, were fixed accordingly in `specs.json`.

This fixes #1122, but note that the code does not need to fetch the datatracker
page for the time being.