Throttle filename check
GitHub did not like the switch to Node.js v20 in jobs. Build fails at the
determine filename step, without much info on top of a "fetch failed".
This update forces throttling on the fetch requests in the hope that this
can improve things.

It also logs specs being processed to provide more clues about what the problem
might be.