Automate processing of IETF drafts - Take 2
The code now fetches all the information it needs for IETF drafts and RFCs from
the IETF datatracker using the Simplified Documents API:
https://datatracker.ietf.org/api/#simplified-documents

This makes it possible to retrieve the latest revision of a document to build
the nightly URL, and to fetch information about the group that standardizes the
document, if any.

IETF documents may be linked to a group, an area, or be part of what IETF calls
individual submissions. Areas and individual submissions still link to a "group"
page at IETF, so the code just takes that info from datatracker as-is. As a
result, individual submissions are no longer associated with the author who
submitted the document, but that does not seem needed in any case.

The code throws when an IETF document that it knows under a certain name got
published under a different name to alert us that the canonical URL needs to
change in browser-specs. Name changes typically happen when a document
transitions to a working group, or when it gets published as an RFC.