Automate processing of IETF documents (#1135)
The code now recognizes IETF draft documents that have a `datatracker.ietf.org`
URL and fetch all the information it needs for IETF drafts and RFCs from the IETF
datatracker using the Simplified Documents API:
https://datatracker.ietf.org/api/#simplified-documents

- It associates them with the IETF organization
- It can compute a useful shortname (that code can in theory return a truncated
shortname because there is no direct way to validate that the Internet Draft
name contains a group ID).
- It extracts the group's from the datatracker API
- It associates IETF documents from the HTTP WG to the right repository.
- It computes the better-looking nightly URL at `www.ietf.org` or at
`httpwg.org` for HTTP WG documents.

This allows to simplify IETF data in `specs.json` a bit.

This fixes #1122, but note that the code does not need to fetch the datatracker
page for the time being.

IETF documents may be linked to a group, an area, or be part of what IETF calls
individual submissions. Areas and individual submissions still link to a "group"
page at IETF, so the code just takes that info from datatracker as-is. As a
result, individual submissions are no longer associated with the author who
submitted the document, but that does not seem needed in any case.

The code throws when an IETF document that it knows under a certain name got
published under a different name to alert us that the canonical URL needs to
change in browser-specs. Name changes typically happen when a document
transitions to a working group, or when it gets published as an RFC.