Fix build when previous build steps already exist (#1137)
Before it runs a build step, the code loads the results of the previous step.
If the file exists, it also loads the previous results of the current step, for
use as fallback if some network issue occurs.

The loading was done using `require()`. The problem is that `require()`
maintains a cache of loaded modules, so if the previous results of the current
step existed, the code actually reused them for the *following* step as well.

This bug did not occur in production because the build job starts from scratch.
In particular, it does not have previous results of build steps.