[tests] More generic rules for repository/sourcePath tests
This fixes #781, using a more generic set of rules. For `repository`, the test
does not look at IETF and FIDO Alliance specs and also skips W3C specs that got
published under www.w3.org and for which we don't have an Editor's Draft.

For `sourcePath`, the test simply skips over entries that don't have a
`repository` property. There remains one exception to the rule but that seems
fine.