Use last published information for discontinued specs (#1140)
This fixes #1130 by turning the "discontinued" standing into a frozen state:
when the last published version of web-specs says that the standing of a spec
is "discontinued", the code directly reuses the information from that last
published version for that spec.

There remained a couple of places in the build script that did not give
priority to information specified in `specs.json`. This update also fixes that
(and needs it!).

If a discontinued spec needs to be revived, the only way to do that will be to
make the new standing explicit in `specs.json`.

Loading the last published version of web-specs could prove useful for #61 as
well, but that's not part of this update.