Better support AOM specs: spec status and PDF release
AOM specs will now be flagged with an AOM status, which can be one of
"Draft Deliverable" or "Final Deliverable". There is unfortunately no
good way to detect that an AOM is a "Final Deliverable", this has to be
specified in `specs.json`.

The code now also automatically creates a `release` property when the
canonical URL of an AOM spec is a PDF.

Now, as far as I can tell, the main AV1 spec is the only one that points at its
PDF version as being the authoritative representation. Other final deliverables
do not say anything about the published PDF. I'm sticking to the HTML URLs as a
result.