Add AOM specs (#1141)
AVIF had already been added to the list. This adds remaining specs from the
Alliance for Open Media related to AV1.

The film grain synthesis spec builds on the same feature *in* AV1, but is
codec-agnostic. Not clear about browser support, so not listing it in the
browser category for now.

Logic also updated to:
- Find more info about AOM specs automatically.
- Assign "Draft Deliverable" or "Final Deliverable" status to AOM specs
- Create a "Release" for AOM specs when the canonical URL is a PDF doc
- Skip "(Draft)" and "(vX.Y)" when computing the short title

Fixes #1088.