Fix handling of CSS specs and CSS snapshots
The W3C API cannot be trusted for CSS specs and CSS snapshots because it
mixes the two in the same series, is case-insensitive on the shortname, and
may change the case of the returned shortname when a new spec gets published in
one of these series. That typically happened with the recent publication of the
CSS 2023 snapshot, which broke the build as a result.

This update makes the code ignore the info from the W3C API for these series,
and rather hardcodes the series titles. The code will also use the latest
version as the current specification for each of these series. If we need to
override this, we'll have to do it in `specs.json`.