Add WASM Core 2 spec and phase 3+ WASM proposals
The list only contained Level 1 of WASM Core (see #1059). This adds Level 2. In
practice, crawling will work fine for the Editor's Draft but Reffy will
basically not extract anything from the /TR version of the spec because it does
not follow usual patterns. That seems fine enough for now.

This update also introduces missing WASM proposals as forks of the WASM Core or
WASM JavaScript API specification. Most of the time, the title needs to be
created because the actual specification remains that of the base spec.

One difficulty is that the WebAssembly group approaches extensions proposals as
generic WASM extensions, and not necessarily as WASM *Core* extensions or WASM
*JS API* extensions, whereas we need to make a choice to set the `forkOf`
property. Things look good for this batch of updates, because current proposals
still seem to extend either of these specs. That may not always be the case
though in the future!

Similarly, the find-specs script assumed that proposals were always extending
the Core spec. It now reports the URL of the home page, both to make it clearer
that a choice needs to be made, and to avoid reporting a proposal that is
already in the list as a WASM JS API fork.

Via #1186.